requirements.txt
flask>=3.0.0
openai>=1.30.0
pypdf>=4.0.0
Pillow>=10.0.0
werkzeug>=3.0.0
gunicorn>=21.0.0
